+++ test/test_pmd_report_builder.rb
@@ -14,7 +14,7 @@ class TestPmdReportBuilder < Test::Unit::TestCase
     record_expecations_after_build
 
     PmdTester::PmdReportBuilder
-      .new(options.base_config, projects, options.local_git_repo, options.base_branch)
+      .new(projects, options, options.base_config, options.base_branch)
       .build
   end
 
@@ -33,7 +33,7 @@ class TestPmdReportBuilder < Test::Unit::TestCase
     record_expecations_after_build
 
     PmdTester::PmdReportBuilder
-      .new(options.base_config, projects, options.local_git_repo, options.base_branch)
+      .new(projects, options, options.base_config, options.base_branch)
       .build
   end
 
@@ -52,7 +52,7 @@ class TestPmdReportBuilder < Test::Unit::TestCase
     record_expectations_project_build
 
     PmdTester::PmdReportBuilder
-      .new(options.base_config, projects, options.local_git_repo, options.base_branch)
+      .new(projects, options, options.base_config, options.base_branch)
       .build
 
     expected = File.read('test/resources/pmd_report_builder/expected-config.xml')
@@ -60,16 +60,37 @@ class TestPmdReportBuilder < Test::Unit::TestCase
     assert_equal(expected, actual)
   end
 
+  def test_build_error_recovery
+    project_list = 'test/resources/pmd_report_builder/project-list.xml'
+    projects = PmdTester::ProjectsParser.new.parse(project_list)
+    assert_equal(1, projects.size)
+    argv = %w[-r target/repositories/pmd -b master -p pmd_releases/6.1.0
+              -c config/design.xml --debug --error-recovery -l]
+    argv.push project_list
+    options = PmdTester::Options.new(argv)
+
+    projects[0].auxclasspath = '-auxclasspath extra:dirs'
+    record_expectations('sha1abc', 'sha1abc', true)
+    record_expecations_after_build
+    record_expectations_project_build(true)
+
+    PmdTester::PmdReportBuilder
+      .new(projects, options, options.base_config, options.base_branch)
+      .build
+  end
+
   private
 
-  def record_expectations_project_build
+  def record_expectations_project_build(error = false)
     PmdTester::ProjectBuilder.any_instance.stubs(:clone_projects).once
     PmdTester::ProjectBuilder.any_instance.stubs(:build_projects).once
     PmdTester::SimpleProgressLogger.any_instance.stubs(:start).once
     PmdTester::SimpleProgressLogger.any_instance.stubs(:stop).once
     File.stubs(:exist?).with('target/reports/master/checkstyle/pmd_report.xml').returns(false).once
+    error_prefix = error ? 'PMD_JAVA_OPTS="-Dpmd.error_recovery -ea" ' : ''
     PmdTester::Cmd.stubs(:execute)
-                  .with('target/pmd-bin-6.10.0-SNAPSHOT/bin/run.sh ' \
+                  .with("#{error_prefix}" \
+                        'target/pmd-bin-6.10.0-SNAPSHOT/bin/run.sh ' \
                         'pmd -d target/repositories/checkstyle -f xml ' \
                         '-R target/reports/master/checkstyle/config.xml ' \
                         '-r target/reports/master/checkstyle/pmd_report.xml ' \